refactor test & until folder
