--- const/index.ts
@@ -1,7 +1,0 @@
-// export const IP = "172.16.50.14:19530";
-// export const IP = '172.16.50.11:19530';
-export const IP = '127.0.0.1:19530';
-
-export const DIMENSION = 4;
-export const INDEX_FILE_SIZE = 1024;
-export const PARTITION_TAG = 'random';
