+++ test/Index.spec.ts
@@ -1,12 +1,12 @@
 import { MilvusClient, ErrorCode } from '../milvus';
-import { IP } from '../const';
 import {
+  IP,
   genCollectionParams,
   VECTOR_FIELD_NAME,
   INDEX_NAME,
   GENERATE_NAME,
-} from '../utils/test';
-import { timeoutTest } from './common/timeout';
+} from './tools';
+import { timeoutTest } from ''./tools'/timeout';
 
 const milvusClient = new MilvusClient({ address: IP, debug: false });
 // names
