+++ test/MilvusClient.spec.ts
@@ -1,6 +1,6 @@
 import { MilvusClient, ERROR_REASONS } from '../milvus';
 import sdkInfo from '../sdk.json';
-import { IP } from '../const';
+import { IP } from './tools';
 
 const milvusClient = new MilvusClient({ address: IP, debug: false });
