+++ test/Replica.spec.ts
@@ -1,10 +1,10 @@
 import { MilvusClient, ERROR_REASONS, ErrorCode } from '../milvus';
-import { IP } from '../const';
 import {
+  IP,
   genCollectionParams,
   VECTOR_FIELD_NAME,
   GENERATE_NAME,
-} from '../utils/test';
+} from './tools';
 
 const milvusClient = new MilvusClient({ address: IP });
 const COLLECTION_NAME = GENERATE_NAME();
