+++ test/tools/index.ts
@@ -1,7 +1,21 @@
-import { DataType, FieldType, convertToDataType } from '../milvus';
+import { DataType, FieldType, convertToDataType } from '../../milvus';
 
+export const IP = '127.0.0.1:19530';
 export const VECTOR_FIELD_NAME = 'vector_field';
 export const INDEX_NAME = 'index_name';
+export const DIMENSION = 4;
+export const INDEX_FILE_SIZE = 1024;
+export const PARTITION_TAG = 'random';
+
+export const timeoutTest = (func: Function, args?: { [x: string]: any }) => {
+  return async () => {
+    try {
+      await func({ ...(args || {}), timeout: 1 });
+    } catch (error) {
+      expect(error.toString()).toContain('DEADLINE_EXCEEDED');
+    }
+  };
+};
 
 /**
  * Generates collection parameters with default fields for a given collection name, dimension, vector type, and optional fields array.
